requirements.txt
aiohttp
openai
python-dotenv
matplotlib
